Se creo y se vinculo la accion de eliminar tareas
