+++ src/components/Task.tsx
@@ -3,12 +3,13 @@ import { Text, View, TouchableOpacity, StyleSheet } from 'react-native'
 import Animated, { useAnimatedStyle, useSharedValue, withTiming } from 'react-native-reanimated';
 import Icon from 'react-native-vector-icons/MaterialIcons';
 import { useAppDispatch } from '../hooks/redux';
-import { ChangueStatus } from '../redux/task/thunk';
+import { ChangueStatus, DeleteTask } from '../redux/task/thunk';
 
 //Componente que contiene la informacion de una tarea, ademas de las acciones posibles a realizar
 export const TaskComponent = ({ task, index, isNewItem }: { task: Task, index: number, isNewItem: MutableRefObject<boolean> }) => {
     const dispatch = useAppDispatch()
     const [checked, setChecked] = useState(task.status); // Estado de la tarea
+
     const toggleCheckbox = () => {// Funcion para cambiar el estado de la tarea
         let newStatus = !checked
         setChecked(newStatus);
@@ -16,6 +17,10 @@ export const TaskComponent = ({ task, index, isNewItem }: { task: Task, index: n
             dispatch(ChangueStatus(task.id, newStatus))
         }, 500);
     };
+
+    const handleDelete = () => {
+        dispatch(DeleteTask(task.id))
+    }
 
     const scale = useSharedValue(index === 0 && isNewItem.current ? 0.5 : 1);
     const opacity = useSharedValue(index === 0 && isNewItem.current ? 0 : 1);
@@ -40,13 +45,13 @@ export const TaskComponent = ({ task, index, isNewItem }: { task: Task, index: n
         <Animated.View style={[style.container, animatedStyles]}>
             <TouchableOpacity onPress={toggleCheckbox} style={{ flexDirection: 'row', alignItems: 'center', flex: 1 }}>
                 <Icon name={checked ? 'check-box' : 'check-box-outline-blank'} size={24} color="black" />
-                <Text style={[{ marginLeft: 8, fontSize: 20, color: "#000000" }, checked && style.checkedText]}>{task.name}</Text>
+                <Text style={[style.titleTask, checked && style.checkedText]}>{task.name}</Text>
             </TouchableOpacity>
             <View style={style.buttonsContainer}>
                 <TouchableOpacity onPress={() => console.log('Editar')} style={style.button}>
                     <Icon name={'edit'} size={24} color="#268cff" />
                 </TouchableOpacity>
-                <TouchableOpacity onPress={() => console.log('Eliminar')} style={style.button}>
+                <TouchableOpacity onPress={handleDelete} style={style.button}>
                     <Icon name={'delete'} size={24} color="#b80c0c" />
                 </TouchableOpacity>
 
@@ -74,4 +79,7 @@ const style = StyleSheet.create({
     checkedText: {
         textDecorationLine: 'line-through',
     },
+    titleTask: {
+        marginLeft: 8, fontSize: 20, color: "#000000", paddingRight: 16
+    }
 })
