+++ src/navigation/Drawer.tsx
@@ -2,7 +2,7 @@ import { createDrawerNavigator } from '@react-navigation/drawer';
 import React from 'react'
 import { NavigationContainer } from '@react-navigation/native';
 import { ToDoList } from '../views/ToDoList';
-import { DeletedList } from '../views/DeletedList';
+import { DoneList } from '../views/DoneList';
 
 export const Drawer = () => {
     const Drawer = createDrawerNavigator();
@@ -10,8 +10,7 @@ export const Drawer = () => {
         <NavigationContainer>
             <Drawer.Navigator>
                 <Drawer.Screen name="List" component={ToDoList} />
-                <Drawer.Screen name="Deleted items" component={DeletedList} />
-
+                <Drawer.Screen name="Done List" component={DoneList} />
             </Drawer.Navigator>
         </NavigationContainer>
     )
