--- src/views/DeletedList.tsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import { Text, View } from 'react-native'
-
-export const DeletedList = () => {
-    return (
-        <View>
-            <Text>
-                Deleted List
-            </Text>
-        </View>
-    )
-}
